requirements.txt
astroid==2.2.5
blis==0.2.4
certifi==2019.11.28
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
colorama==0.4.1
cymem==2.0.3
idna==2.8
isort==4.3.21
joblib==0.14.1
lazy-object-proxy==1.4.1
mathparse==0.1.2
mccabe==0.6.1
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.4
pandas==0.25.3
Pillow==7.0.0
Pint==0.9
plac==0.9.6
preshed==2.0.1
pylint==2.3.1
pymongo==3.10.0
pytesseract==0.3.2
python-dateutil==2.7.5
pytz==2019.3
PyYAML==3.13
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
spacy==2.1.9
SQLAlchemy==1.2.19
srsly==0.2.0
tesseract==0.1.3
thinc==7.0.8
tqdm==4.40.2
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.9
wasabi==0.4.2
wrapt==1.11.2
